Fix collect function to group results properly
I was using groupBy without first sorting, which was causing incorrect
groupings of the results. Changed to use the general groupResultsBy
fuction that does the proper grouping.

This bug only showed up on linux when expanding a directory due to the
order in which the files were read from the directory.