Get pin scripts working with new tool locations
The pin tools have been moved to a new location and the scripts
needed to be updated to deal with these locations. We also made the
scripts report failures more reliably by using die and close on the
file handles. Finally, we made the scripts work for both 64-bit and
32-bit machines.